fix typo

+++ shader/fragmentShader.glsl
@@ -8,8 +8,8 @@ void main(void)
 {
 	float x = screenPosition.x;
 	float y = screenPosition.y;
-	int m = 1;
-	int n = 2;
+	float m = 1;
+	float n = 2;
 	float z = abs(cos(n * PI * x) * cos(m * PI * y) - cos(m * PI * x) * cos(n * PI * y));
 	gl_FragColor = vec4(z, z, z, 1.0);
 }
